kapm-backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-decouple==3.8
Pillow==11.3.0
drf-spectacular==0.28.0
django-filter==24.3
whitenoise==6.9.0
celery==5.4.0
redis==5.0.8
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3